refactor: new helper convention for useNow
