+++ components/schedule/ScheduleAtAGlance.tsx
@@ -2,12 +2,12 @@
 import HoverList from "@/components/person/HoverList";
 import GlanceCountdown from "@/components/schedule/GlanceCountdown";
 import cn from "@/utils/helpers/cn";
-import { useNow } from "@/utils/helpers/date";
 import {
   getCurrentPeriod,
   getTodaySetToPeriodTime,
 } from "@/utils/helpers/schedule";
 import { getLocaleString } from "@/utils/helpers/string";
+import useNow from "@/utils/helpers/useNow";
 import { useLocale } from "@/utils/hooks/i18n";
 import { StylableFC } from "@/utils/types/common";
 import { UserRole } from "@/utils/types/person";
